Player file 2018-19 (Not complete)
Mostly an experimental version of a 2018-19 player file. It might never get done, but I've used it to figure out how to make a player file for next season. I imported the data from a BBState CSV that gave me the names, was able to convert heights, move over weights, convert positions and school names are all there too (though BBstate has the mascots and that's bad for this file) and so there's still a lot of work for the utility for this year, but would be a few weeks of work to get a file for 2019-20 together if I wanted to do that. With help, could be even faster -- the hardest part will be player ratings - and I haven't even conceptualized that part yet.